reintroduce combined ReplyAction
